fix(api-load): Allow false for unload
Currently, only `true` and an array of strings are valid values for the property. While `false` does not add additional functionality, it removes an odd and unexpected behavior of triggering unload in the same form as if you were to pass `true`. Other than specifically not unloading when `false`, this change avoids opinionation of the value passed itself.

Ref #321