perf: more efficient shape offset calculation
`getShapeOffset` drives the shape calculations for nearly every chart type. It also did a lot of unnecessary looping over data points. In stacked charts, this was especially obvious as it meant it was re-calculating the stacking many extra times.

The changes here move the data around smarter so that re-calculation doesn't occur, significantly increasing speed.