chore(build): build optimization
1) Import the ESM version of babel helpers
This allows webpack to concatenate them with other modules during
optimizations, instead of having to keep them in separate scopes.

2) Disable the webpack polyfilling of node APIs

The library has a code path using global to get global variables, in
some cases. But it already has its own handling of other environment and
so does not need the webpack polyfill. And the webpack polyfill prevents
applying module concatenation on that module.

Fix #1081
Close #1096