chore(types): Document that the callback options are called without this
Previous types would make typescript think that callbacks are called
with this being the Options object (or some subpart of it), and so will
not report errors related to it. By specifying the type of this as void,
we tell the type checker that callbacks must not use `this`.
The implementation is actually calling some callbacks with the
ChartInternal object as context, but this is considered an
implementation detail coming from the fork of C3, not as a supported
feature.

Fix #1246
Close #1295