feat(gauge): added support for drawing gauge from 'startingAngle' to 'arcLength'
- updated gauge config options
- 'arcLength' is limited to: -100 <== arcLength (in percent) <== 100
- values <-100 and >100 will default to -100 or 100

- If 'gauge_fullCircle: false':
-- 'startingAngle >== -1 * Math.PI / 2' defaults to -1 * Math.PI / 2
-- 'startingAngle <== Math.PI / 2' defaults to Math.PI / 2
- If 'gauge_fullCircle: true':
-- 'startingAngle >== -1 * Math.PI' defaults to -1 * Math.PI
-- 'startingAngle <== Math.PI' defaults to Math.PI

If gauge is not full circled, there is no reason to use starting angle lower
than our default and greater than the opposite of our default, as it would
draw the arc out of the svg.
If gauge is full circled, there is no reason to use starting angle lower
than -PI and greater than PI as all relevant starting angles are already
in this range. There are no benefits for accepting other values and it would
make the calculations for the text placements harder.

updateRadius() has to be called to make sure,
'gaugeArcWidth' is calculated correctly before
drawing arcs

Close #1633
Close #1803
Close #1849