fix(axis): removed state mismatch due to wrong conditions in 'getMaxTickWidth'
There was a missing check if every 'currentTicksMax.domain' is greater than zero.
Due to this, if new data was laoded with 'unload: true' it could happen, that
'currentTicksMax.domain' would be [0,0].
In this case, even if the domain is not same, it would always result in returning
'currentTicksMax.size'and not updating 'currentTicksMax.domain' which resulted
in not updating the rotation on x axis.