fix(module): fix module compatibility
Downgrade d3.js dependencies from v3 to v2,
where the adoption of type:module breaks compatibility
in different envrionment(ex. Next.js)

Is due to type:module force to be ESM, but most of the
ecosystem environments aren't ready for this now.

To maintain the compatibility, downgrade is needed for now.

Ref #2181